Changing the payoffs of this extensive form
The payoffs were wrong and didn't respect the rules of the game. Here is the corrected version:
***** Red part (left) *****
- Raises 3 : If Benoît meets Mahsa bet, as the card is red, Mahsa wins everything on the table so she wins 7 (initial bet of 4 + raise 3 of Benoit) and Benoit loses everything -> +7/-7 . If Benoit passes, then he loses only his initial bet, and Mahsa wins it : +4/-4 (Mahsa keeps its bet + wins the 4 of Benoit)
- Folds: Mahsa shows her card, as it is red she wins all the money on the table, so she wins Benoit's bet and Benoit loses its bet : +4/-4
- Raises 4 : same reasoning as for the raise of 3 but with 4 now: 8/-8 for meeting and 4/-4 for passing
***** Black part (left) *****
Exact same reasoning but as the card is black, when Mahsa folds in this case it is Benoit that wins everything (-4/4-, when Benoit passes (in both case of raises) then it is the same payoffs as the card is not shown and it is only Benoit that loses its initial bet (4/-4 in each cases), and for the two meets, Benoit wins everything because the card is black (-7/7 for raising 3 and -8/8 for raising 4).