Create 0702.tex
I have written the results obtained in the exercise session on Latex, with some explanation. I am just not sure how the sequential equilibrium needs to be written, I guess that's only one sequential equilibrium taking account of the randomised strategy + d_2 but I don't know what's the correct notation. For the other points, I have tried to explain the best that I could... Tell me if I need to change anything! 

Maybe a tikzpicutre for the last point could be helpful but when I tried to add the lambda, beta, and so on it did not looked good